Update README.md
Miles/Kilometers

Write a program that will convert miles to kilometers and kilometers to miles. The user will indicate both a number (representing a distance) and a choice of whether that number is in miles to be converted to kilometers or kilometers to be converted to miles. Each conversion is done with a value returning function. You may use the following conversions.

1 kilometer = .621 miles
1 mile = 1.61 kilometers

Sample Run 1:

Please input 
  1 Convert miles to kilometers 
  2 Convert kilometers to miles 
1
Please input the miles to be converted 
120 
120 miles = 193.2 kilometers 

Sample Run 2:

Please input 
  1 Convert miles to kilometers 
  2 Convert kilometers to miles 
2 
Please input the kilometers to be converted 
235 
235 kilometers = 145.935 miles 

Optional Step

Include a loop that lets the user repeat this computation for new input values until the user says they want to end the program.

Sample Run:

Please input 
  1 Convert miles to kilometers 
  2 Convert kilometers to miles 
  3 Quit 
1
Please input the miles to be converted 
120 
120 miles = 193.2 kilometers 
Please input 
  1 Convert miles to kilometers 
  2 Convert kilometers to miles 
  3 Quit 
2 
Please input the kilometers to be converted 
235 
235 kilometers = 145.935 miles 
Please input 
  1 Convert miles to kilometers 
  2 Convert kilometers to miles 
  3 Quit 
3